Bad design for the scene to handle the player moves - instead keyevents handled by the player who has focus. Goal keepers and computer team not allowed to acquire focus.
Needs more testing.

+++ trunk/goalkeeper.cpp
@@ -83,8 +83,8 @@ void GoalKeeper::gkAdvanceWithoutBall()
             || (dir == Team::NorthToSouth
                 && m_pitch->m_topPenaltyArea->contains(m_pitch->ball()->pos())) ) {
 
-            int dx = abs(pos().x() - m_pitch->ball()->pos().x());
-            int dy = abs(pos().y() - m_pitch->ball()->pos().y());
+            qreal dx = abs(pos().x() - m_pitch->ball()->pos().x());
+            qreal dy = abs(pos().y() - m_pitch->ball()->pos().y());
             if ( m_pitch->ball()->controlledBy() && ( dx < 5) && (dy < 5) )
                 action = MWindow::Tackle;
             else
